% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallLeaf.R
\name{CanopyWindS}
\alias{CanopyWindS}
\title{Calculates canopy wind shelter coefficient}
\usage{
CanopyWindS(h, pai)
}
\arguments{
\item{h}{canopy height (m)}

\item{pai}{a vector of plant area index values as returned by [PAIgeometry()]}
}
\value{
a vector of of the same length as `pai` of canopy wind shelter coefficients.
}
\description{
Generates a vector of the same length as `pai` of coefficients by
which to multiply wind speed at the top of the canopy
}
\examples{
# plant area within each layer
pai1 <- PAIgeometry(5, 7, 70)
pai2 <- PAIgeometry(5, 5, 50)
z <- c(1:1000) / 100
uh <- 3 # Wind speed at top of canopy
uz1 <- uh * CanopyWindS(10, pai1)
uz2 <- uh * CanopyWindS(10, pai2)
plot(z ~ uz1, type = "l", col = "blue",
     xlab = "Wind speed", xlim = c(0, 3), lwd = 2)
par(new = TRUE)
plot(z ~ uz2, type = "l", col = "red",
     xlab = "", xlim = c(0, 3), lwd = 2)
}
