% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ground.R
\name{creategroundp}
\alias{creategroundp}
\title{Ground parameters from soil types}
\usage{
creategroundp(soiltype)
}
\arguments{
\item{soiltype}{a character vector of soil type - one of those listed in [soilparams()]}
}
\value{
object of class `groundparams`
}
\description{
Generates an object of class `groundparams` from a specified
soil type.
}
\examples{
creategroundp("Loam")
}
