% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewMicroPoint.R
\name{createvegp}
\alias{createvegp}
\title{Create list of vegetation parameters for running the microclimate model}
\usage{
createvegp(vegtype = "BET.Te")
}
\arguments{
\item{vegtype}{Character string giving the plant function type for which
parameters are to be created (see details)}
}
\value{
a list of vegetation parameters:
\describe{
  \item{h}{Vegetation height (m)}
  \item{pai}{Plant Area Index (m^2 m^-2)}
  \item{x}{Campbell leaf angle distribution parameter (unitless)}
  \item{lref}{Leaf reflectance (shortwave) (unitless)}
  \item{ltra}{Leaf transmittance (shortwave) (unitless)}
  \item{lrefp}{Leaf reflectance (PAR) (unitless)}
  \item{ltrap}{Leaf transmittance (PAR) (unitless)}
  \item{Vcmx25}{Maximum Rubisco carboxylation rate at 25 °C (µmol m^-2 s^-1)}
  \item{Tup}{Upper temperature limit for photosynthesis (°C)}
  \item{Tlow}{Lower temperature limit for photosynthesis (°C)}
  \item{Dcrit}{Vapour pressure deficit at onset of strong stomatal limitation (kPa)}
  \item{alpha}{Photosynthetic quantum efficiency (mol CO2 mol^-1 PAR)}
  \item{Kxmx}{Maximum xylem hydraulic conductance (µmol m^-2 s^-1)}
  \item{hv}{Huber value (cm^2 m^-2)}
  \item{f0}{Ratio of ci/(ci − Γ*) under low vapour pressure deficit from Jacobs (1994) (unitless)}
  \item{fd}{Fraction of Vcmax converted to leaf dark respiration (unitless)}
  \item{psi50}{Water potential causing 50\% loss of maximum hydraulic conductance (MPa)}
  \item{apsi}{Xylem hydraulic conductance parameter (calculated if \code{< 0}) (unitless)}
  \item{len}{Leaf length (m)}
  \item{wid}{Leaf width (m)}
  \item{vegem}{Vegetation emissivity (unitless)}
  \item{mwft}{Maximum water film thickness (mm)}
  \item{pTAW}{Fraction of total available soil water that can be depleted before stress (unitless)}
  \item{rootskew}{Root distribution skew towards the top of the soil profile (unitless)}
}
}
\description{
Create list of vegetation parameters for running the microclimate model
}
\details{
Recognised plant functional types are `BET.Tr`	- Broadleaf Evergreen Tree (Tropical),
`BET.Te` - Broadleaf Evergreen Tree (Temperate), 	`BDT`	-  Broadleaf Deciduous Tree,
`NET` - Needleleaf Evergreen Tree,	`NDT`	- Needleleaf Deciduous Tree, `C3` - C3 grass,
	`C4` - C4 grass,	`ESh` - Evergreen Shrub and	`DSh`- Deciduous Shrub.
}
