% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigLeaf.R
\name{.BigLeaf}
\alias{.BigLeaf}
\title{Run big-leaf model: one iteration}
\usage{
.BigLeaf(
  weather,
  zref,
  vegp,
  groundp,
  bigleafp,
  solar,
  twostreamp,
  soilm,
  lat,
  long,
  merid,
  dst,
  gmn,
  i,
  method,
  yearG
)
}
\description{
Run big-leaf model: one iteration
}
