% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiation.R
\name{.cank2}
\alias{.cank2}
\title{Internal function for calculating canopy extinction coefficients for sloped ground surfaces}
\usage{
.cank2(zen, x, si)
}
\description{
Internal function for calculating canopy extinction coefficients for sloped ground surfaces
}
