% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallLeaf.R
\name{.CanopyHL}
\alias{.CanopyHL}
\title{Canopy sensible heat flux}
\usage{
.CanopyHL(
  weather,
  forestp,
  solar,
  twostreamp,
  smallleafp,
  tground,
  groundem = 0.97,
  i,
  surfwet = 0.98
)
}
\description{
Canopy sensible heat flux
}
