% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallLeaf.R
\name{.LangrangianOne}
\alias{.LangrangianOne}
\title{Run one iteration of Langrangian below canopy model}
\usage{
.LangrangianOne(
  i,
  weather,
  forestp,
  solar,
  twostreamp,
  smallleafp,
  groundem = 0.97,
  surfwet = 1,
  theta = 1,
  a0 = 0.25,
  a1 = 1.25
)
}
\description{
Run one iteration of Langrangian below canopy model
}
