% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallLeaf.R
\name{.RunSmallLeafOne}
\alias{.RunSmallLeafOne}
\title{Run one step of small leaf model}
\usage{
.RunSmallLeafOne(
  i,
  weather,
  forestp,
  solar,
  twostreamp,
  smallleafp,
  lat,
  long,
  groundem = 0.97,
  surfwet = 1,
  theta = 1,
  a0 = 0.25,
  a1 = 1.25,
  iter = 10,
  bwgt = 0.75
)
}
\description{
Run one step of small leaf model
}
