% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallLeaf.R
\name{.SmallLeafAs}
\alias{.SmallLeafAs}
\title{Assign values to smallleafp for timestep i}
\usage{
.SmallLeafAs(smallleafp, th, eh, uh, h, d, zm, ws, bigleafp, n, i)
}
\description{
Assign values to smallleafp for timestep i
}
