% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snow.R
\name{.steadystate}
\alias{.steadystate}
\title{Derives steady-state temperature and energy balance}
\usage{
.steadystate(Rabs, gHa, tc, pk, rh, G = 0)
}
\description{
Derives steady-state temperature and energy balance
}
