% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigLeaf.R
\name{.stomcond}
\alias{.stomcond}
\title{Stomatal conductance (set g50 to 300 and multiply by 3 to give bulk surface)}
\usage{
.stomcond(Rsw, gsmax, q50 = 100)
}
\description{
Stomatal conductance (set g50 to 300 and multiply by 3 to give bulk surface)
}
