% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runfunctions.R
\name{dtr_correct}
\alias{dtr_correct}
\title{apply diurnal temperature range correction to climate data}
\usage{
dtr_correct(climdata, zref, uzref = zref, lat, long)
}
\arguments{
\item{climdata}{a data.frame of hourly weather as in the example dataset
`climdata` (see details)}

\item{zref}{height above ground (m) of e.g. temperature measurements in `climdata`}

\item{lat}{latitude (decimal degrees). Negative south of the equator.}

\item{long}{longitude (decimal degrees). Negative west of the Greenwich Meridian.}

\item{uref}{height above ground (m) of wind speed measurements in `climdata` (if different from zref)}

\item{details}{Climate datasets derived from e.g. ERA5 or NCEP/NOAA often
underestimate the diurnal temperature range in coastal areas, where a grid
cell is classified as sea. When running the microclimate model in areas with
tall vegetation, an automatic height adjustment is performed to derive estimates
of wind speed, temperature and relative humidity at the top of the canopy. This
is achieved by assuming land surface properties typical of a weather station and
extrapolating the vertical profiles upwards. However, in instances where the original
climate dataset significantly underestimates the diurnal temperature range, this
can cause problems. When extrapolated upward, the temperature profiles become
too variable and hence a warning is given. This function corrects diurnal temperatures ranges by adopting
the following rational: temperatures at greater height will always will always
have a lower diurnal temperature range than at lower heights as the diurnal cycle
is caused by absorption of radiation by the ground surface. The diurnal temperature
range is incrementally adjusted until this is achieved.}
}
\description{
The function `dtr_correct` applies a correction to temperature
and relative humidity to enable more realistic height adjustments
}
