% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewMicroPoint.R
\name{expand_output}
\alias{expand_output}
\title{R wrapper for c++ expand_outputCpp function}
\usage{
expand_output(mat, nout)
}
\arguments{
\item{mat}{matrix to expand.}

\item{nout}{number of rows of expanded matrix}
}
\value{
an expanded matrix
}
\description{
Spline interpolates a Numeric Matrix for better visual representation
}
