% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forestparams}
\alias{forestparams}
\title{A dataset of vegetation parameters for running the Small Leaf model.}
\format{
A list of  the following objects:
\describe{
  \item{h}{vegetation height (m)}
  \item{pai}{plant area index}
  \item{x}{ratio of vertical to horizontal projections of leaf foliage}
  \item{clump}{canopy clumping factor quantified as fraction of direct radiation passing through larger gaps in the canopy when the sun is at its zenith}
  \item{lref}{leaf reflectance}
  \item{ltra}{leaf transmittance}
  \item{leafd}{leaf diameter (m)}
  \item{em}{emissvity of vegetation}
  \item{gsmax}{maximum stomatal conductances (mol / m^2 / s)}
  \item{q50}{value of PAR when stomatal conductances is at 50 percent of its maximum value (micro mol/m^2/s)}
  \item{skew}{degree of skew towards top of canopy in foliage}
  \item{spread}{degree of spread in foliage}
}
}
\usage{
forestparams
}
\description{
An object of class forestparams
}
\keyword{datasets}
