% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ectotherm.R
\name{full_ecto}
\alias{full_ecto}
\title{Run the full ectotherm model through time}
\usage{
full_ecto(
  mout,
  climdata,
  vegp,
  soilc,
  ectop,
  lat,
  long,
  maxiter = 100,
  tolerance = 0.01
)
}
\arguments{
\item{mout}{List of microclimate variables as returned by [RunModelFull()]}

\item{climdata}{data.frame of meteorological forcing data using the same
naming conventions as the inbuilt dataset `climdata`.}

\item{vegp}{Vegetation parameter list as returned by [createvegp()].}

\item{soilc}{Soil parameter list as returned by [createsoilc()].}

\item{ectop}{List of ectotherm parameters required by [Ectotherm()].}

\item{lat}{Numeric. Latitude in decimal degrees.}

\item{long}{Numeric. Longitude in decimal degrees.}

\item{maxiter}{Integer. Maximum number of iterations used to achieve
convergence in the ectotherm model.}

\item{tolerance}{Numeric. Convergence tolerance for the ectotherm model.}
}
\value{
A list of two Numeric Matrices:
\describe{
  \item{Tbodya}{Estimated body temperature above ground (degrees C).}
  \item{Tbody}{Estimated body temperature below ground (degrees C).}
}
}
\description{
Runs the full ectotherm model for all time steps and heights using microclimate
data returned by [RunModelFull()]
}
\details{
Below ground, body temperature is assumed to equal soil temperature.
}
