% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ground.R
\name{GFlux}
\alias{GFlux}
\title{Computes ground heat flux}
\usage{
GFlux(Tg, soilm, rho, Vm, Vq, Mc, Gmax = 200, Gmin = 200, i = 1, yearG = TRUE)
}
\arguments{
\item{Tg}{a vector of ground surface temperatures (deg C)}

\item{soilm}{a vector of fractional soil moistures}

\item{rho}{Soil bulk density (Mg / m^3)}

\item{Vm}{Volumetric mineral fraction of soil}

\item{Vq}{Volumetric quartz fraction of soil}

\item{Mc}{Mass fraction of clay}

\item{Gmax}{Maximum allowable value of daily G (included to ensure model convergence - W/m^2)}

\item{Gmin}{Minumum allowable value of daily G (included to ensure model convergence - W/m^2)}

\item{i}{iteration of model}

\item{yearG}{optional logical indicating whether or not to calculate and incorporate yearly ground heat flux}
}
\value{
a list of the following:
(1) G - ground heat fluxes (W/m^2)
(2) Gmax - newly compued values of Gmax if on first iteration of model
(2) Gmin - newly compued values of Gmax if on first iteration of model
}
\description{
Computes ground heat flux from hourly surface temperature data
}
\details{
The maximum allowable values of Gmax and Gmin are computed on the first
iteration of the model in which they are derived from the diurnal soil surface temperature
values with G set to 0. It is thus assumed that G can only dampen rather than enhance soil
surface temperature fluctuations,an assumption that increases the likelihood of iterative
convergence
}
