% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groundparams}
\alias{groundparams}
\title{A dataset of ground parameters.}
\format{
A list of  the following objects:
\describe{
  \item{gref}{ground reflectance}
  \item{slope}{slope of ground surface (deg)}
  \item{aspect}{aspect of ground surface (deg from north)}
  \item{em}{emissvity of gorund surface}
  \item{rho}{Soil bulk density (Mg / m^3)}
  \item{Vm}{Volumetric mineral fraction of soil}
  \item{Vq}{Volumetric quartz fraction of soil}
  \item{Mc}{Mass fraction of clay}
  \item{b}{Shape parameter for Campbell soil moisture model}
  \item{Psie}{Matric potential (J / m^3)}
  \item{Smax}{Volumetric water content at saturation}
  \item{Smin}{Residual water content}
  \item{alpha}{Shape parameter of the van Genuchten model (cm^-1)}
  \item{n}{Pore size distribution parameter}
  \item{Ksat}{Saturated hydraulic conductivity (cm / day)}
}
}
\usage{
groundparams
}
\description{
An object of class groundparams
}
\keyword{datasets}
