% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlantGeometry.R
\name{PAIgrass}
\alias{PAIgrass}
\title{Generates plant area index profile for grass habitat}
\usage{
PAIgrass(hgt, n = 1000, PAI = NA, fraccover = NA, taper = 1)
}
\arguments{
\item{hgt}{grass height (m) - of tallest grass spike (~1.5x drop disk measured height)}

\item{n}{Number of plant area index values to generate. Default: 1000 (see details)}

\item{PAI}{Optionally total plant area index of canopy.}

\item{fraccover}{Optionally grass fractional cover}

\item{taper}{parameter controlling how tapered the plant area index
is towards the top. A value of one is typical of grass. Lower values give
a more even spread. Cannot be negative or zero.}
}
\value{
a vector of length `n` of plant area index values the sum of which equals `PAI`
}
\description{
Generates a vector of length `n` of plausible plant area index values
}
\details{
if `PAI` is not supplied it is estimated from `fraccover`, which must
be between 0.001 and 0.9999. If neither `PAI` or `fraccover` is supplied, `PAI`
is estimated using a simple allometric relationship with sward height,
derived from field measurements from a meadow in northern Spain.
}
\examples{
paii <- PAIgrass(hgt = 0.25, n = 1000)
folden <- paii * 1000
z <- (c(1:1000) / 1000) * 0.25
plot(z ~ folden, type = "l", main = sum(paii)) # plots foliage density
}
