% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runfunctions.R
\name{plotprofile}
\alias{plotprofile}
\title{plot temperature or humidity profile}
\usage{
plotprofile(
  climdata,
  hr,
  plotout = "tair",
  vegp,
  paii = NA,
  groundp,
  lat,
  long,
  zref = 2,
  uref = zref,
  soilm = NA,
  surfwet = NA,
  dTmx = 25,
  maxiter = 20,
  n = 100
)
}
\arguments{
\item{climdata}{a data.frame of hourly weather as in the example dataset
`climdata` (see details)}

\item{hr}{the hour in climdata for which the height profile is to be plotted}

\item{plotout}{of `tair`, `tleaf`, `ea` or `relhum`}

\item{vegp}{an object of class vegparams or forestparams formatted as for the inbuilt example
datasets `vegparams` or `forestparams` (see details).}

\item{paii}{optional vector of plant area index values for each canopy layer
as returned by [PAIgeometry()] (see details).}

\item{groundp}{an object of class groundparams formatted as for the inbuilt
example dataset `groundparams`.}

\item{lat}{latitude (decimal degrees). Negative south of the equator.}

\item{long}{longitude (decimal degrees). Negative west of the Greenwich Meridian.}

\item{zref}{height above ground (m) of measurements in `weather` (see details)}

\item{uref}{height above gorund (m) of wind speed measurements in `climdata` (if different from zref)}

\item{soilm}{optional vector of hourly volumetric soil moisture fractions in
the top 10 cm of soil. Calculated if not supplied.}

\item{surfwet}{optional vector of canopy surface acting like a saturated
water surface. Calculated if not supplied.}

\item{dTmx}{= maximum by which vegetation or soil surface temperature can
exceed air temperature (deg C, set to ensure model convergence).}

\item{maxiter}{= maximum number of iterations over which to run the model to
achieve model convergence.}

\item{n}{number of canopy layers used to plot the model. if `paii` supplied,
ignored and taken from length of `paii`.}
}
\value{
a list of the following
\describe{
 \item{z}{height above ground (m)}
 \item{var}{one of leaf temperature (degrees C, air temperature (degrees C),
 vapour pressure (kPa) or relative humidity (percentage) for eahc height `z`}
}
`soilm` are returned.
}
\description{
The function `plotprofile` plots a temperature or humidity height
profile and returns a list of heights or humidites.
above or below ground.
}
\details{
if `plotout` = `tleaf` the height propofile of leaf tmeperature is plotted
from  the ground surface to the top of the canopy. If `plotout` = `tair`, `ea` or `relhum`
the air temperature (tair), vapour pressure (ea) or relative humidity (relhum) profiles
are plotted from the ground to two metres above canopy.
}
\examples{
# plot air temperature for hottest hour
xx<-plotprofile(climdata, hr = 4094, "tair", forestparams, paii = NA, groundparams,
  lat =49.96807, long= -5.215668)
# plot leaf temperature for hottest hour
xx<-plotprofile(climdata, hr = 4094, "tleaf", forestparams, paii = NA, groundparams,
  lat =49.96807, long= -5.215668)
}
