% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snow.R
\name{pointsnow}
\alias{pointsnow}
\title{RUns a simple snow model}
\usage{
pointsnow(
  weather,
  vegp,
  initdepth = 0.01,
  inittemp = 0,
  snowenv = "Taiga",
  zref = 2
)
}
\arguments{
\item{weather}{a data.frame of hourly weather as in the example dataset `climdata`}

\item{vegp}{an object of class vegparams formatted as for the inbuilt example dataset `vegparams`}

\item{initdepth}{snow temperature (deg C) at start of model run}

\item{snowenv}{either one of 'Tundra', 'Taiga', 'Prairie', 'Alpine' or 'Maritime'
used for estimating snow density as function of snowpack age following the exponential
function of Sturm et al. 2010 J. of Hydromet. 11:1380-1394. Alternatively a vector of
parameters for the Sturm et al function.}

\item{zref}{height above ground of measurements in `weather`. Adjust using [WeatherHeight()]
if `zref` < `vegp$h`.}
}
\value{
a list of the following:
\describe{
\item{tme}{POSIXlt object of times and dates corresponding to model outputs}
 \item{snowdepth}{a vector of snow depths (m)}
 \item{snowtemp}{a vector of snow temperatures (deg C)}
 \item{snowalb}{a vector of snow albedos (0-1)}
 \item{snowdens}{a vector of snow densities (kg/m^3)}
 \item{psim}{diabatic correction factors for heat}
 \item{psih}{diabatic correction factors for momentum}
}
}
\description{
Calculates snow depth and temperature
}
\examples{
snowm <- pointsnow(snowclimdata, vegparams)
plot(snowm$snowdepth ~ as.POSIXct(snowclimdata$obs_time), type = "l",
     xlab = "Date", ylab = "Depth (m)")
}
