% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiation.R
\name{RadiationBigLeaf}
\alias{RadiationBigLeaf}
\title{Calculates radiation fluxes}
\usage{
RadiationBigLeaf(
  weather,
  vegp,
  groundp,
  bigleafp,
  solar,
  twostreamp,
  lat = 49.97,
  long = 5.22,
  merid = 0,
  dst = 0
)
}
\arguments{
\item{weather}{a data.frame of hourly weather as in the example dataset `climdata`}

\item{vegp}{an object of class vegparams formatted as for the inbuilt example dataset `vegparams`}

\item{groundp}{an object of class vegparams formatted as for the inbuilt example dataset `groundparams`}

\item{bigleafp}{a list with ground and canopy heat exchange surface temperature as returned by [RunBigLeaf()]}

\item{solar}{a list of zenith and azimuth angles as returned by [solarposition()]}

\item{twostreamp}{a list of two-stream radiation model parameters as returned by [twostreamparams()]}

\item{lat}{latitude (decimal degrees). Negative south of the equator.}

\item{long}{latitude (decimal degrees). Negative west of the Greenwich Meridian.}

\item{merid}{Longitude of local time zone meridian (decimnal degrees), Default: 0 (Greenwich Mean Time)}

\item{dst}{Daylight saving hours. E.g. with `merid = 0` 0 for UTC or 1  for British Summer Time.}
}
\value{
A list of the following:
(1) radGsw - flux density of shortwave radiation absorbed by canopy 'big-leaf' (W/m^2)
(2) radGlw - flux density of longwave radiation absorbed by canopy 'big-leaf'(W/m^2)
(3) radCsw - flux density of shortwave radiation absorbed by ground surface (W/m^2)
(4) radClw - flux density of shortwave radiation absorbed by ground surface (W/m^2)
}
\description{
Calculates radiation fluxes for application with BigLeaf model
}
\details{
Outputs form this function are used to calculate and the ground and canopy heat
exchange surfaces. The outputs `radGsw` and `radGlw` are for the entire 'bigleaf'.
I.e. include both the canopy and the ground surface.
}
