% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiation.R
\name{RadiationSmallLeaf}
\alias{RadiationSmallLeaf}
\title{Calculates shortwave radiation fluxes within canopy}
\usage{
RadiationSmallLeaf(
  weather,
  forestp,
  solar,
  twostreamp,
  twostreamPAR,
  smallleafp,
  tground,
  groundem = 0.97,
  i
)
}
\arguments{
\item{weather}{a data.frame of hourly weather as in the example dataset `climdata`}

\item{forestp}{an object of class vegparams formatted as for the inbuilt example dataset `vegparams`}

\item{solar}{a list of zenith and azimuth angles as returned by [solarposition()]}

\item{twostreamp}{a list of two-stream radiation model parameters as returned by [twostreamparams()]}

\item{twostreamPAR}{a list of two-stream radiation model parameters for photosynthetically
active radiation (PAR) as returned by [twostreamparams()] but with leaf reflectance
and transmittance adjusted for PAR.}

\item{smallleafp}{an object of class smallleafp as returned by [RunBigLeaf()]}

\item{tground}{ground temperature at time increment i (deg C)}

\item{groundem}{ground emissivity}

\item{i}{entry of weather for which shortwave fluxes are required}
}
\value{
A list of the following:
(1) swupper - flux density of shortwave radiation absorbed by upper-side of canopy elements (W/m^2)
(2) swunder - flux density of shortwave radiation absorbed by under-side of canopy elements (W/m^2)
(3) lwupper - flux density of longwave radiation absorbed by upper-side of canopy elements (W/m^2)
(4) lwunder - flux density of longwave radiation absorbed by under-side of canopy elements (W/m^2)
(5) PARupper - Downward flux density of PAR  (W/m^2)
(6) PARunder - Upward flux density of PAR (W/m^2)
}
\description{
Calculates radiation fluxes for application with SmallLeaf model
}
