% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigLeaf.R
\name{RunBigLeaf}
\alias{RunBigLeaf}
\title{Run big-leaf model}
\usage{
RunBigLeaf(
  climdata,
  vegp,
  groundp,
  lat,
  long,
  zref = 2,
  uref = zref,
  soilm = NA,
  surfwet = NA,
  dTmx = 25,
  maxiter = 20
)
}
\arguments{
\item{vegp}{an object of class vegparams formatted as for the inbuilt example dataset `vegparams`}

\item{groundp}{an object of class vegparams formatted as for the inbuilt example dataset `groundparams`}

\item{lat}{latitude (decimal degrees). Negative south of the equator.}

\item{long}{longitude (decimal degrees). Negative west of the Greenwich Meridian.}

\item{zref}{height above ground (m) of measurements in `weather` (see details)}

\item{uref}{height above gorund (m) of wind speed measurements in `climdata` (if different from zref)}

\item{soilm}{a vector of volumetric soil moisture fractions in the top 10 cm of soil.}

\item{dTmx}{maximum by which vegetation or soil surface temperature can exceed air temperature (deg C, set to ensure model convergence)}

\item{maxiter}{Maximum number of iterations over which to run the model}

\item{weather}{a data.frame of hourly weather as in the example dataset `climdata`}

\item{merid}{Longitude of local time zone meridian (decimnal degrees), Default: 0 (Greenwich Mean Time)}

\item{dst}{Daylight saving hours. E.g. with `merid = 0` 0 for UTC or 1  for British Summer Time.}

\item{bwgt}{backward weighting to apply when iteratively running the model (default 0.5)}

\item{tol}{Error margin (deg C) deemed acceptable for model convergence (default 0.5).}

\item{gmn}{optional minimum convective conductance value (mol/m^2/s). Lower values are more
physically realistic, but can reduce likelihood of model convergence.}

\item{plotout}{optional logical indicating whether to plot a time-series of ground surface
temperatures on each iteration}

\item{swmethod}{method used to calculate fraction of ground surface acting as free water surface
(0 = based on rainfall, 1 computed from soil effective relative humidity, 2 computed from soil moisture fraction)}

\item{yearG}{optional logical indicating whether or not to calculate and account for annual ground heat flux cycle}
}
\value{
an object of class pointmicro, namely a list of the following:
\describe{
 \item{tme}{POSIXlt object of times and dates corresponding to model outputs}
 \item{Tc}{a vector of canopy heat exchange surface temperatures (deg C)}
 \item{Tg}{a vector of ground surface temperatures (deg C)}
 \item{H}{a vector of sensible heat fluxes (W/m^2)}
 \item{G}{a vector of ground heat fluxes (W/m^2)}
 \item{psih}{diabatic correction factors for heat}
 \item{psim}{diabatic correction factors for momentum}
 \item{phih}{diabatic influencing factors for heat}
 \item{OL}{Obukhov length}
 \item{error.mar}{maximum temperature difference between ultimate and penultimate iteration of model}
}
}
\description{
Runs big leaf model iteratively to derive ground and canopy
heat exchange surface temperatures
}
\details{
When running the model `zref` must be greater than he height of vegetation.
If running for a location with tall vegetation, function [WeatherHeight()] can
be used to adjust e.g. temperature and wind speed to a user-specified height.
}
\examples{
# Run Big leaf model with inbuilt parameters
bigleafp <- RunBigLeaf(climdata, vegparams, groundparams, lat = 49.96807,
long = -5.215668)
# Plot ground and canopy temperature
tme <- as.POSIXct(climdata$obs_time, tz = "UTC")
par(mar = c(6, 6, 3, 3))
# Ground temperature
plot(bigleafp$Tg ~ tme, type = "l", cex.axis = 2, cex.lab = 2,
     xlab = "Month", ylab = "Temperature", ylim = c(-5, 40),
     col = rgb(1, 0, 0, 0.5))
par(new = TRUE)
# Canopy temperature
plot(bigleafp$Tc ~ tme, type = "l", cex.axis = 2, cex.lab = 2,
     xlab = "", ylab = "", ylim = c(-5, 40), col = rgb(0, 0.5, 0, 0.5))
}
