% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runfunctions.R
\name{runpointmodel}
\alias{runpointmodel}
\title{runs point microclimate model}
\usage{
runpointmodel(
  climdata,
  reqhgt,
  vegp,
  paii = NA,
  groundp,
  lat,
  long,
  zref = 2,
  uref = zref,
  soilm = NA,
  surfwet = NA,
  dTmx = 25,
  maxiter = 20,
  n = 20,
  runchecks = FALSE
)
}
\arguments{
\item{climdata}{a data.frame of hourly weather as in the example dataset
`climdata` (see details)}

\item{reqhgt}{height (m) above (positive) or below (negative) ground for
which model output is wanted.}

\item{vegp}{an object of class vegparams or forestparams formatted as for the inbuilt example
datasets `vegparams` or `forestparams` (see details).}

\item{paii}{optional vector of plant area index values for each canopy layer
as returned by [PAIgeometry()] (see details).}

\item{groundp}{an object of class groundparams formatted as for the inbuilt
example dataset `groundparams`.}

\item{lat}{latitude (decimal degrees). Negative south of the equator.}

\item{long}{longitude (decimal degrees). Negative west of the Greenwich Meridian.}

\item{zref}{height above ground (m) of e.g. temperature measurements in `climdata`}

\item{uref}{height above gorund (m) of wind speed measurements in `climdata` (if different from zref)}

\item{soilm}{optional vector of hourly volumetric soil moisture fractions in
the top 10 cm of soil. Calculated if not supplied.}

\item{surfwet}{optional vector of canopy surface acting like a saturated
water surface. Calculated if not supplied.}

\item{dTmx}{= maximum by which vegetation or soil surface temperature can
exceed air temperature (deg C, set to ensure model convergence).}

\item{maxiter}{= maximum number of iterations over which to run the model to
achieve model convergence.}

\item{n}{number of canopy layers used to run model. if `paii` supplied,
ignored and taken from length of `paii`.}
}
\value{
if `reqhgt > 0` a Dataframe with the following columns:
\describe{
 \item{obs_time}{POSIXlt object of dates and times}
 \item{tair}{Air temperature (degrees C) at height `reqhgt`}
 \item{tleaf}{Leaf temperature (degrees C) at height `reqhgt` or `tcanopy` average canopy temperature if `reqhgt` above canopy}
 \item{relhum}{relative humidity (percentage) at height `reqhgt`}
 \item{windspeed}{wind speed (m/s) at height `reqhgt`}
 \item{Rdirdown}{Total downward flux of direct radiation (W / m^2) at height `reqhgt` perpendicular to solar beam}
 \item{Rdifdown}{Total downward flux of diffuse radiation (W / m^2) at height `reqhgt`}
 \item{Rlwdown}{Total downward flux of longwave radiation (W / m^2) at height `reqhgt`}
 \item{Rswup}{Total upward flux of shortwave radiation (W / m^2) at height `reqhgt` (assumed diffuse)}
 \item{Rlwup}{Total upward flux of longwave radiation (W / m^2) at height `reqhgt`}
}
if `reqhgt = 0` `tair` is replaced by `tground` representing ground surface
temperature (deg C), `relhum` by `soilm` (fractional soil moisture) and `tleaf`
and `windspeed` are not returned. If `reqhgt < 0` only `obs_time`, `tground` and
`soilm` are returned.
}
\description{
The function `runpointmodel` runs the point microclimate model
above or below ground.
}
\details{
Variable `obs_time` in `climdata` assumes times are in UTC. To derive
within canopy temperatures , the canopy is divided into `n` layers
and each element `i` of `paii` represents the plant area index within a layer at
height `(i / n) * hgt` where `n` is the total number of layers and `hgt` is canopy
height. If `paii` is not supplied, `vegp` must contain entries for `skew` and
`spread` as used by [PAIgeomtry()], which is used to generate a plausable
vector of `paii` values. Note that `sum(paii)` is assumed  to equal `vegp$pai`,
the later representing the total leaf area of the canopy.
}
\examples{
# run model using inbuilt datasets and defaults
mout<-runpointmodel(climdata, reqhgt = 1, forestparams, paii = NA, groundparams,
  lat =49.96807, long= -5.215668)
plot(mout$tair,type="l")
}
