% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallLeaf.R
\name{RunSmallLeaf}
\alias{RunSmallLeaf}
\title{Runs small leaf model to derive below canopy temperatures}
\usage{
RunSmallLeaf(
  weather,
  forestp,
  groundp,
  reqhgt,
  lat,
  long,
  zref = 2,
  n = 10,
  soilm = NA,
  dTmx = 25,
  merid = 0,
  dst = 0,
  maxiter = 100,
  smalliter = 10,
  bwgt = 0.5,
  sbwgt = 0.5,
  tol = 0.5,
  gmn = 0.1,
  a0 = 0.25,
  a1 = 1.25,
  plotout = FALSE,
  swmethod = 2,
  yearG = TRUE,
  saveprofile = FALSE
)
}
\arguments{
\item{weather}{a data.frame of hourly weather as in the example dataset `climdata`}

\item{forestp}{an object of class forestparams formatted as for the inbuilt example dataset `forestparams`}

\item{groundp}{an object of class vegparams formatted as for the inbuilt example dataset `groundparams`}

\item{reqhgt}{height (m) above ground for which microclimate variables are required}

\item{lat}{latitude (decimal degrees). Negative south of the equator.}

\item{long}{longitude (decimal degrees). Negative west of the Greenwich Meridian.}

\item{zref}{height above ground (m) of measurements in `weather`}

\item{n}{number of layers to divide canopy into (higher = more accurate but slower). Default 100.}

\item{soilm}{a vector of volumetric soil moisture fractions in the top 10 cm of soil. Calculated if not supplied.}

\item{dTmx}{maximum by which vegetation or soil surface temperature can exceed air temperature (deg C, set to ensure model convergence)}

\item{merid}{Longitude of local time zone meridian (decimnal degrees), Default: 0 (Greenwich Mean Time)}

\item{dst}{Daylight saving hours. E.g. with `merid = 0` 0 for UTC or 1  for British Summer Time.}

\item{maxiter}{Maximum number of iterations over which to run the big leaf model}

\item{smalliter}{number of iterations over which to run the small leaf model (default 10)}

\item{bwgt}{backward weighting to apply when iteratively running the big leaf model model (default 0.5)}

\item{sbwgt}{backward weighting to apply when iteratively running the small leaf model model (default 0.5)}

\item{tol}{Error margin (deg C) deemed acceptable for big leaf model model convergence (default 0.5).}

\item{gmn}{optional minimum convective conductance value (mol/m^2/s) for big leaf model. Lower values are more
physically realistic, but can reduce likelihood of model convergence.}

\item{a0}{thermal diffusivity scale parameter from Raupach 1989 Agr Forest Meteorol, 47: 85-108.}

\item{a1}{thermal diffusivity shape parameter from Raupach 1989 Agr Forest Meteorol, 47: 85-108.}

\item{plotout}{optional logical indicating whether to plot vertical profile of air temperatures every 12 hours (default FALSE)
temperatures on each iteration}

\item{swmethod}{method used to calculate fraction of ground surface acting as free water surface in big leaf model
(0 = based on rainfall, 1 computed from soil effective relative humidity, 2 computed from soil moisture fraction)}

\item{yearG}{optional logical indicating whether or not to calculate and account for annual ground heat flux cycle in big leaf model}

\item{saveprofile}{optional logical indicating whether to return microclimate variables for all canopy layers (TRUE) or
or just for height `reqhgt` (FALSE). Default FALSE.}
}
\value{
a a list of the following:
\describe{
 \item{tme}{POSIXlt object of times and dates corresponding to model outputs}
 \item{tair}{a vector or matrix of air temperatures (deg C)}
 \item{tleaf}{a vector or matrix of leaf temperatures (deg C)}
 \item{relhum}{a vector or matrix of relative humidities (Percentage)}
 \item{windspeed}{a vector or matrix of wind speeds (m/s)}
 \item{Rdni_down}{a vector or matrix of downward direct normal irradiance fluxes (W/m^2)}
 \item{Rdif_down}{a vector or matrix of downward diffuse radiation fluxes (W/m^2)}
 \item{Rsw_up}{a vector or matrix of upward shortwave radiation fluxes (W/m^2), assumed entirely diffuse}
 \item{Rlw_down}{a vector or matrix of downward longward radiation fluxes (W/m^2)}
 \item{Rlw_up}{a vector or matrix of upward longwave radiation fluxes (W/m^2)}
}
}
\description{
Runs small leaf model for each time-step in turn to derive below
canopy temperature, humidity and radiation fluxes.
}
\details{
If `saveprofile = TRUE` the returned values are matrices, with rows corresponding
to each canopy layer and columns corresponding to each time step. If `saveprofile = FALSE`
returned values are vectors for height `reqhgt`, with elements corresponding to each time
step.
}
\examples{
# Select first 5 days of May from inbuilt climate dataset
tme <- as.POSIXlt(climdata$obs_time, tz = "UTC")
sel <- which(tme$mon + 1 == 5 & tme$mday < 6)
weather <- climdata[sel,]
# Run model with inbuilt vegetation and ground parameters and default settings
smallleafp <- RunSmallLeaf(weather, forestparams, groundparams, reqhgt = 5,
                           lat = 49.96807, long = -5.215668, yearG = FALSE)
# Plot air temperature at reqhgt
plot(smallleafp$tair ~ as.POSIXct(tme[sel]), type = "l",
     xlab = "Day", ylab = "Temperature")
}
