% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallLeaf.R
\name{RunSmallLeafOne}
\alias{RunSmallLeafOne}
\title{Run one step of small leaf model}
\usage{
RunSmallLeafOne(
  i,
  weather,
  forestp,
  twostreamp,
  smallleafp,
  tground,
  lat,
  long,
  groundem = 0.97,
  surfwet = 1,
  theta = 1,
  a0 = 0.25,
  a1 = 1.25,
  iter = 10
)
}
\arguments{
\item{i}{time increment of model}

\item{weather}{a data.frame of hourly weather as in the example dataset `climdata`}

\item{forestp}{an object of class vegparams formatted as for the inbuilt example dataset `forestparams`}

\item{twostreamp}{parameters of two-stream radiation model as returned by [twostreamparams()]}

\item{smallleafp}{a list returned by [RunSmallLeafOne()] in previous time increment}

\item{tground}{ground surface temperature (deg C)}

\item{lat}{latitude (decimal degrees). Negative south of the equator.}

\item{long}{longitude (decimal degrees). Negative west of the Greenwich Meridian.}

\item{groundem}{ground emissivity temperature (deg C)}

\item{surfwet}{fraction of canopy surface acting like a saturated evaporating surface}

\item{theta}{fraction of ground surface acting like a saturated evaporating surface}

\item{a0}{optional coefficient controlling vertical variation in eddy variance}

\item{a1}{optional coefficient controlling vertical variation in eddy variance}

\item{iter}{number of iterations over which to run model to stablise outputs}
}
\value{
A list of the following:
(1) tleaf - a vector representing vertical variation in leaf temperatures (deg C)
(2) tair - a vector representing vertical variation in air temperatures a vector representing vertical variation in leaf temperatures (deg C)
(3) pai - a vector representing vertical variation in plant area index values
(4) pai_a - a vector representing vertical variation in plant area index values above each point of interest
(5) uz - a vector representing vertical variation in wind speed (m/s)
(6) ea - a vector representing vertical variation in vapour pressure (kPa)
(7) psim - diabatic correction coefficient for momentum at canopy top
(8) psih - diabatic correction coefficient for heat at canopy top
(9) phih - diabatic influencing factor at canopy top
(10) lwgt - a list of weightings used for calculating longwave radiation
}
\description{
Runs a single step of the small leaf model
}
