% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ground.R
\name{soilmmodel}
\alias{soilmmodel}
\title{Runs a simple two-layer soil model}
\usage{
soilmmodel(weather, soiltype)
}
\arguments{
\item{weather}{a data.frame of hourly weather formatted as [climdata()]}

\item{soiltype}{a character vector of soil type - one of those listed in [soilparams()]}
}
\value{
a vector of hourly volumetrivc soil moisture fractions
}
\description{
Runs a simple two-layer soil model to derive soil moisture in upper 10 cm
}
\examples{
soilm<-soilmmodel(climdata, "Loam")
plot(soilm,type="l")
}
