% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soilparams}
\alias{soilparams}
\title{A dataset of soilparameters parameters.}
\format{
A data frame with the following columns:
\describe{
  \item{Soil.type}{description of soil type}
  \item{Smax}{Volumetric water content at saturation (m^3 / m^3)}
  \item{Smin}{Residual water content (m^3 / m^3)}
  \item{alpha}{Shape parameter of the van Genuchten model (cm^-1)}
  \item{n}{Pore size distribution parameter (dimensionless, > 1)}
  \item{Ksat}{Saturated hydraulic conductivity (cm / day)}
  \item{Vq}{Volumetric quartz fraction of soil}
  \item{Vm}{Volumetric mineral fraction of soil}
  \item{Vo}{Volumetric organic fraction of soil}
  \item{Mc}{Mass fraction of clay}
  \item{rho}{Soil bulk density (Mg / m^3)}
  \item{b}{Shape parameter for Campbell model (dimensionless, > 1)}
  \item{psi_e}{Matric potential (J / m^3)}
  \item{mult}{soil moisture model radiation coefficient}
  \item{rmu}{soil moisture model rainfall coefficient}
  \item{a}{soil moisture model deeper layer multiplier coefficient}
  \item{pwr}{soil moisture model deeper layer power coefficient}
}
}
\source{
: \url{https://onlinelibrary.wiley.com/doi/full/10.1002/ird.1751}
}
\usage{
soilparams
}
\description{
An object of class groundparams
}
\keyword{datasets}
