% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiation.R
\name{solarposition}
\alias{solarposition}
\title{Calculates solar position}
\usage{
solarposition(lat, long, year, month, day, hour, merid = 0, dst = 0)
}
\arguments{
\item{lat}{latitude (decimal degrees)}

\item{long}{longitude(decimal degrees, -ve west of Greenwich meridian)}

\item{year}{year or POSIXlt object of times}

\item{month}{month. Ignored if year is a POSIXlt object}

\item{day}{day of month. Ignored if year is a POSIXlt object}

\item{hour}{decimal hour. Ignored if year is a POSIXlt object}

\item{merid}{meridian of local time zone (0 for UTC)}

\item{dst}{daylight saving yours (e.g. one for British Summer Time)}
}
\value{
a list of zenith and azimuth angles (decimal degrees)
}
\description{
Calculates the solar zenith and azimuth angles
}
