% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ectotherm.R
\name{timeseries_ecto}
\alias{timeseries_ecto}
\title{Return time-series of ectotherm body-temperature for a specified height}
\usage{
timeseries_ecto(
  mout,
  climdata,
  reqhgt,
  vegp,
  soilc,
  ectop,
  lat,
  long,
  maxiter = 100,
  tolerance = 0.01
)
}
\arguments{
\item{mout}{List of microclimate variables as returned by [RunMicro()]}

\item{climdata}{data.frame of meteorological forcing data using the same
naming conventions as the inbuilt dataset `climdata`.}

\item{reqhgt}{height above or below ground for which body temperature is wanted
(negative is below ground). Must match value supplied to [RunMIcro()].}

\item{vegp}{Vegetation parameter list as returned by [createvegp()].}

\item{soilc}{Soil parameter list as returned by [createsoilc()].}

\item{ectop}{List of ectotherm parameters required by [Ectotherm()].}

\item{lat}{Numeric. Latitude in decimal degrees.}

\item{long}{Numeric. Longitude in decimal degrees.}

\item{maxiter}{Integer. Maximum number of iterations used to achieve
convergence in the ectotherm model.}

\item{tolerance}{Numeric. Convergence tolerance for the ectotherm model.}
}
\value{
A data.frame with two columns:
\describe{
  \item{obs_time}{POSIXlt object of dates and times associated with eahc body
  temperature estimate}
  \item{Tbody}{Predicted ectotherm body temperature (degrees C).}
}
}
\description{
Runs the ectotherm model through time using microclimate data returned by
[RunMicro()] and returns predicted body temperature for very hour.
}
\details{
If `confrac` in `ectop` is > 0, and `0 < reqhgt < vegp$h` the ectotherm is
assumed to be sitting on leaves. If `reqhgt = 0` the ectotherm is assumed
to be sitting on the ground. Below ground, body temperature is
assumed to equal soil temperature.
}
