% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiation.R
\name{twostreamparams}
\alias{twostreamparams}
\title{Calculates two-stream radiation model parameters}
\usage{
twostreamparams(vegp, groundp, solar)
}
\arguments{
\item{vegp}{an object of class vegparams formatted as for the inbuilt example dataset `vegparams`}

\item{groundp}{an object of class vegparams formatted as for the inbuilt example dataset `groundparams`}

\item{solar}{a list of zenith and azimuth angles as returned by [solarposition()]}
}
\value{
a list of parameters for two-stream approximation equations
}
\description{
Calculates parameters of Dickinson-Sellers two-stream
radiation model
}
\details{
For details of the Dickinson-Sellers two-stream radiation
model see Sellers (1985) https://doi.org/10.1080/01431168508948283.
The function implements improvements to the model proposed by Yuan et
al (2017) https://doi.org/10.1002/2016MS000773
}
\examples{
# Run two-sream model with inbuilt datasets
tme <- as.POSIXlt(climdata$obs_time, tz="UTC")
# POSIXlt object used in place of year:
solar <- solarposition(49.96807, -5.215668, year=tme)
twostreamp <- twostreamparams(vegparams, groundparams, solar)
# Calculate white- and black-sky albedo
cl <- vegparams$clump
Kc <- with(twostreamp, kkd$kd / kkd$k0)
albd <- with(twostreamp, p1 + p2 + cl^2 * groundparams$gref)
albb <- with(twostreamp, p5 / sig + p6 + p7 + cl^Kc * gref2)
# Calculate and plot blue-sky albedo
alb <- with(climdata, (albd * difrad + albb * (swdown - difrad)) / swdown)
# Set blue-sky albedo to be white-sky albedo when direct radiation is zero
dni<-with(climdata, swdown - difrad)
alb[dni == 0] <- albd
par(mar = c(6, 6, 3, 3))
plot(alb ~ as.POSIXct(tme), type="l", cex.axis = 2, cex.lab = 2,
     xlab = "Month", ylab = "Albedo", ylim = c(0, 1))
}
