% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vegparams}
\alias{vegparams}
\title{A dataset of vegetation parameters for running Big Leaf model.}
\format{
A data.frame with the following elements:
\describe{
  \item{h}{vegetation height (m)}
  \item{pai}{plant area index}
  \item{x}{ratio of vertical to horizontal projections of leaf foliage}
  \item{clump}{canopy clumping factor quantified as fraction of direct radiation passing through larger gaps in the canopy when the sun is at its zenith}
  \item{lref}{leaf reflectance}
  \item{ltra}{leaf transmittance}
  \item{leafd}{leaf diameter (m)}
  \item{em}{emissvity of vegetation}
  \item{gsmax}{maximum stomatal conductances (mol / m^2 / s)}
  \item{q50}{value of PAR when stomatal conductances is at 50 percent of its maximum value (micro mol/m^2/s)}
}
}
\usage{
vegparams
}
\description{
A data.frame of parameters for running the soil models
}
\keyword{datasets}
