% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigLeaf.R
\name{weather_hgt}
\alias{weather_hgt}
\title{Adjust weather data for height}
\usage{
weather_hgt(climdata, zin = 2, uzin = zin, zout = 10, lat, long)
}
\arguments{
\item{zin}{height above ground of temperature and relative humidity measurements (m)}

\item{uzin}{height above ground of wind speed measurements (m)}

\item{zout}{height above ground for which observations are needed}

\item{lat}{latitude (decimal degrees). Negative south of the equator.}

\item{long}{latitude (decimal degrees). Negative west of the Greenwich Meridian.}

\item{weather}{a data.frame of hourly weather as in the example dataset `climdata`}
}
\value{
a data.frame of hourly weather
}
\description{
Adjust weather data to user-specified height above ground
}
\details{
When running [RunBigLeaf()], `zref` the height at which weather
observations are assumed to have been taken must be greater than the
height of vegetation. This function derives estimates of temperature,
relative humidity and wind speed at a user specified height above ground,
by assuming that weather observations were taken in an open area with
with a vegetation height of 12 cm and a plant area index of 1, consistent
with WMO guidelines for locating weather stations.
}
\examples{
weather<-weather_hgt(climdata, 2, 2, 10, 49.96807, -5.215668)
# Temperature comparison
plot(weather$temp ~ climdata$temp, xlab = "Temperature at 2 m",
     ylab = "Temperature at 10m", pch = 15)
abline(a = 0, b = 1, lwd = 2, col = "red")
# Relative humidity comparison
plot(weather$relhum ~ climdata$relhum, xlab = "Relative humidity at 2 m",
     ylab = "Relative humidity at 10m", pch = 15)
abline(a = 0, b = 1, lwd = 2, col = "red")
# Wind speed comparison
plot(weather$windspeed ~ climdata$windspeed, xlab = "Wind speed at 2 m",
     ylab = "Wind speed at 10m", pch = 15)
abline(a = 0, b = 1, lwd = 2, col = "red")
}
