% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallLeaf.R
\name{vegpforgrass}
\alias{vegpforgrass}
\title{Generates vegetation parameters for grass}
\usage{
vegpforgrass(hgt, n = 20, taper = 1)
}
\arguments{
\item{hgt}{grass height (m) - of tallest grass spike (~1.5x drop disk measured height)}

\item{n}{Number of plant area index values to generate. Default: 20 (see details)}

\item{taper}{parameter controlling how tapered the plant area index
is towards the top. A value of one is typical of grass. Lower values give
a more even spread. Cannot be negative or zero.}
}
\value{
a vector of length `n` of plant area index values the sum of which equals `PAI`
}
\description{
Generates an object of class
`vegparams` and a vector of of length `n` of plausible plant
area index values for grassy vegetation.
}
\details{
a list comrpising the following:
\describe{
 \item{vegp}{An object of vegetation parameters for running the microclimate model}
 \item{paii}{a vector of length `n` of plant area index values}
}
}
\examples{
vp <- vegpforgrass(0.25, 20)
vegp <- vp$vegp
paii <- vp$paii
mout <- runpointmodel(climdata, reqhgt = 0.12, vegp, paii,  groundparams, lat = 49.96807, long= -5.215668)
tme <- as.POSIXct(climdata$obs_time)
plot(mout$tair ~ tme, type="l", xlab = "Month", ylab = "Air temperature",
     ylim = c(-5, 40), col = "red")
}
